services/what-if-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
neo4j==5.15.0
clickhouse-connect==0.6.23
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
structlog==23.2.0
networkx==3.2.1
numpy==1.25.2
scipy==1.11.4